+++ frontend/src/pages/PostPage.tsx
@@ -1,12 +1,11 @@
 import React, {useEffect, useState} from 'react';
-import styled from "styled-components";
-import PostSmall from "large-components/PostSmall/PostSmall";
-import CommentSection from "large-components/CommentSection";
-import {useAppSelector} from "other/hooks";
-import { useParams } from 'react-router';
 import axios from "axios";
-import {IPostMini} from "../other/widelyUsedTypes";
-import {rawDataPostOne} from "../other/mocking-data/firstLoadData";
+import styled from "styled-components";
+import { useParams } from 'react-router';
+import PostSmall from "large-components/PostSmall";
+import CommentSection from "large-components/CommentSection";
+import {IPostMini} from "other/widelyUsedTypes";
+import {rawDataPostOne} from "other/mocking-data/firstLoadData";
 
 function PostPage() {
     const params = useParams();
